create new suitespec
