+++ hatch.toml
@@ -288,12 +288,7 @@ dependencies = [
     "cryptography",
     "astunparse",
     "simplejson",
-    "SQLAlchemy",
-    "psycopg2-binary~=2.9.9",
-    "pymysql",
-    "mysqlclient==2.1.1",
     "grpcio",
-    "mysql-connector-python",
 ]
 
 [envs.appsec_iast_default.env-vars]
@@ -430,6 +425,45 @@ test = [
 
 [[envs.iast_tdd_propagation.matrix]]
 python = ["3.9", "3.10", "3.11", "3.12", "3.13"]
+
+## IAST packages integrations
+
+[envs.appsec_integrations_packages]
+template = "appsec_integrations_packages"
+dependencies = [
+    "pytest",
+    "pytest-cov",
+    "hypothesis",
+    "requests",
+    "urllib3",
+    "pycryptodome",
+    "cryptography",
+    "astunparse",
+    "simplejson",
+    "SQLAlchemy",
+    "psycopg2-binary~=2.9.9",
+    "pymysql",
+    "mysqlclient==2.1.1",
+    "grpcio",
+    "mysql-connector-python",
+]
+
+[envs.appsec_integrations_packages.env-vars]
+_DD_IAST_PATCH_MODULES = "benchmarks.,tests.appsec."
+DD_IAST_REQUEST_SAMPLING = "100"
+DD_IAST_DEDUPLICATION_ENABLED = "false"
+
+
+[envs.appsec_integrations_packages.scripts]
+test = [
+    "uname -a",
+    "pip freeze",
+    "python -m pytest --no-ddtrace {args:tests/appsec/integrations/packages_tests/}",
+]
+
+[[envs.appsec_integrations_packages.matrix]]
+python = ["3.8", "3.9", "3.10", "3.11", "3.12", "3.13"]
+
 
 ## ASM appsec_integrations_django
