+++ tests/appsec/integrations/fixtures/patch_psycopg2.py
@@ -1,0 +1,6 @@
+from psycopg2.extensions import adapt
+
+
+def adapt_list(obj_list):
+    value = adapt(obj_list)
+    return b"r-" + value.getquoted()
