+++ tests/appsec/integrations/fixtures/sql_injection_mysqldb.py
