+++ tests/appsec/integrations/fixtures/sql_injection_psycopg2.py
@@ -1,16 +1,16 @@
-from pymysql.err import OperationalError
+from psycopg2.errors import DuplicateTable
 
 from ddtrace.appsec._iast._taint_tracking._taint_objects import get_tainted_ranges
 from ddtrace.appsec._iast._taint_tracking._taint_objects import is_pyobject_tainted
-from tests.appsec.iast.db_utils import get_pymysql_connection
+from tests.appsec.integrations.packages_tests.db_utils import get_psycopg2_connection
 
 
 def sqli_simple(table):
-    connection = get_pymysql_connection()
+    connection = get_psycopg2_connection()
     cur = connection.cursor()
     try:
         cur.execute("CREATE TABLE students (name TEXT, addr TEXT, city TEXT, pin TEXT)")
-    except OperationalError:
+    except DuplicateTable:
         connection.rollback()
     # label test_sql_injection
     cur.execute("SELECT 1 FROM " + table)
