+++ tests/appsec/integrations/fixtures/sql_injection_sqlite3.py
